requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
django-crispy-forms==1.14.0
django-tempus-dominus==5.1.2.17
sqlparse==0.4.2
tzdata==2022.1
